This is the project based on machine learning
This project is based on the MATLAB platform for experimentation. In the early stage, a large amount of pile foundation low strain testing data from the Anhui Province Construction Science and Research Institute is classified and marked manually. In the middle stage, the low strain reflection curve is tested on site, with the aim of establishing a time series neural network for the automatic classification of pile foundation defects. Ultimately, a model capable of automatically classifying the defect level of pile foundation low strain testing will be completed. This will not only enable rapid, intelligent, and effective judgment of the pile foundation quality level, but it also has great commercial potential.